feat: refactor component imports, add error handling utility, and implement organisation form
