+++ src/App.jsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import SignUp from './page/Admin/Registration/SignUp.jsx'
 import Login from './page/Admin/login/Login.jsx'
-import Organisation from './components/OrganisationForm'
-import Invite from './components/Invite'
+import Organisation from './page/Admin/organization-form/OrganisationForm.jsx'
+import Invite from './page/Admin/invite/Invite.jsx'
 import Dashboard from './page/Admin/Dashboard/Dashboard.jsx'
 import Users from './page/Admin/UserManagement/Users.jsx'
